fixes bad requests and adds middleware

+++ controllers/ingredientCtrl.js
@@ -5,6 +5,7 @@ const allIngredients = async (req, res) => {
     where : {id: req.user.id},
     include: [db.ingredient]
   })
+  
   if (typeof foundUser.ingredients[0] === "undefined") {
     return res.json({status: 404, message: "this user has no ingredients"})
   }
@@ -13,15 +14,9 @@ const allIngredients = async (req, res) => {
 
 const addIngredient = async (req, res) => {
   if (req.body.name !== "" ) {
-    const [ingredient, created] = await db.ingredient.findOrCreate({
-      where: { name: req.body.name }
-    })
+    const ingredient = await db.ingredient.create({name: req.body.name})
     
     const foundUser = await db.user.findOne({ where: { id: req.user.id } })
-    //in case no user is found
-    if (!foundUser) {
-     return res.json({ message: "No user found"})
-    }
   
     foundUser.addIngredient(ingredient)
     .then(relationInfo => {
@@ -39,7 +34,11 @@ const removeIngredient = async (req, res) => {
   const ingredients = foundUser.ingredients
   
   const ingredient = ingredients.filter(el => el.name === req.body.name)
-
+  console.log("ingredient called")
+  if (typeof ingredient[0] === "undefined") {
+    console.log("no ingredient")
+    return res.json({message: "No ingredients of that name"})
+  }
   db.ingredient.destroy({where: {id: ingredient[0].id}})
   .then(data => res.json({message: "Success"}))
   .catch(error => {
